+++ src/lib/djerba/plugins/genomic_landscape/plugin.py
@@ -6,12 +6,13 @@ import csv
 from djerba.plugins.base import plugin_base, DjerbaPluginError
 from mako.lookup import TemplateLookup
 import djerba.core.constants as core_constants
+import djerba.util.oncokb.constants as oncokb_constants
 from djerba.plugins.genomic_landscape.provenance_tools import parse_file_path
 from djerba.plugins.genomic_landscape.provenance_tools import subset_provenance
 from djerba.util.render_mako import mako_renderer
 import djerba.util.input_params_tools as input_params_tools
 from djerba.util.subprocess_runner import subprocess_runner
-from djerba.util.oncokb.annotator import oncokb_annotator
+from djerba.util.oncokb.annotator import annotator_factory
 from djerba.util.oncokb.tools import levels as oncokb_levels
 import djerba.util.oncokb.constants as okb
 import djerba.plugins.genomic_landscape.constants as constants
@@ -28,6 +29,9 @@ class main(plugin_base):
     PLUGIN_VERSION = '1.0.0'
     TEMPLATE_NAME = 'genomic_landscape_template.html'
 
+    # TODO standardize this constant, see ticket GCGI-1290
+    INPUT_PARAMS_ONCOTREE_CODE = 'oncotree_code'
+
     # For MSI file
     MSI_RESULTS_SUFFIX = '.recalibrated.msi.booted'
     MSI_WORKFLOW = 'msisensor'
@@ -41,7 +45,7 @@ class main(plugin_base):
       discovered = [
            constants.DONOR,
            constants.TUMOUR_ID,
-           constants.ONCOTREE_CODE,
+           oncokb_constants.ONCOTREE_CODE,
            constants.TCGA_CODE,
            constants.PURITY_INPUT,
            constants.MSI_FILE,
@@ -50,6 +54,12 @@ class main(plugin_base):
       for key in discovered:
           self.add_ini_discovered(key)
       self.set_ini_default(core_constants.ATTRIBUTES, 'clinical')
+      self.set_ini_default(
+          oncokb_constants.ONCOKB_CACHE,
+          oncokb_constants.DEFAULT_CACHE_PATH
+      )
+      self.set_ini_default(oncokb_constants.APPLY_CACHE, False)
+      self.set_ini_default(oncokb_constants.UPDATE_CACHE, False)
 
       # Default parameters for priorities
       self.set_ini_default('configure_priority', 100)
@@ -64,69 +74,39 @@ class main(plugin_base):
     def configure(self, config):
       config = self.apply_defaults(config)
       wrapper = self.get_config_wrapper(config)
-
-      if wrapper.my_param_is_null(constants.ONCOTREE_CODE):
-          if self.workspace.has_file(input_params_helper.INPUT_PARAMS_FILE):
-              data = self.workspace.read_json(input_params_helper.INPUT_PARAMS_FILE)
-              oncotree_code = data[constants.ONCOTREE_CODE]
-              wrapper.set_my_param(constants.ONCOTREE_CODE, oncotree_code)
-          else:
-              msg = "Cannot find Oncotree code; must be manually specified or "+\
-                    "given in {0}".format(input_params_helper.INPUT_PARAMS_FILE)
-              self.logger.error(msg)
-              raise DjerbaPluginError(msg)
-      
-      if wrapper.my_param_is_null(constants.TCGA_CODE):
-          if self.workspace.has_file(input_params_helper.INPUT_PARAMS_FILE):
-              data = self.workspace.read_json(input_params_helper.INPUT_PARAMS_FILE)
-              tcga_code = data[constants.TCGA_CODE]
-              wrapper.set_my_param(constants.TCGA_CODE, tcga_code)
-          else:
-              msg = "Cannot find TCGA code; must be manually specified or "+\
-                    "given in {0}".format(input_params_helper.INPUT_PARAMS_FILE)
-              self.logger.error(msg)
-              raise DjerbaPluginError(msg)
-
-      if wrapper.my_param_is_null(constants.PURITY_INPUT):
-          if self.workspace.has_file(input_params_helper.INPUT_PARAMS_FILE):
-              data = self.workspace.read_json(input_params_helper.INPUT_PARAMS_FILE)
-              purity = data[constants.PURITY_INPUT]
-              wrapper.set_my_param(constants.PURITY_INPUT, purity)
-          else:
-              msg = "Cannot find Purity; must be manually specified or "+\
-                    "given in {0}".format(input_params_helper.INPUT_PARAMS_FILE)
-              self.logger.error(msg)
-              raise DjerbaPluginError(msg)
-
-      if wrapper.my_param_is_null(constants.DONOR):
-          if self.workspace.has_file(input_params_helper.INPUT_PARAMS_FILE):
-              data = self.workspace.read_json(input_params_helper.INPUT_PARAMS_FILE)
-              donor = data[constants.DONOR]
-              wrapper.set_my_param(constants.DONOR, donor)
-          else:
-              msg = "Cannot find Purity; must be manually specified or "+\
-                    "given in {0}".format(input_params_helper.INPUT_PARAMS_FILE)
-              self.logger.error(msg)
-              raise DjerbaPluginError(msg)
-
+      # mapping of input params JSON keys -> genomic landscape INI keys
+      input_keys = [constants.TCGA_CODE, constants.PURITY_INPUT, constants.DONOR]
+      config_keys = {k:k for k in input_keys}
+      config_keys[self.INPUT_PARAMS_ONCOTREE_CODE] = oncokb_constants.ONCOTREE_CODE
+      # try to find config values
+      for k in config_keys.keys():
+          if wrapper.my_param_is_null(config_keys[k]):
+              if self.workspace.has_file(input_params_helper.INPUT_PARAMS_FILE):
+                  data = self.workspace.read_json(input_params_helper.INPUT_PARAMS_FILE)
+                  wrapper.set_my_param(config_keys[k], data[k])
+                  self.logger.debug("Read {0} from input params JSON".format(k))
+              else:
+                  msg = "Cannot find {0}; must be manually specified or ".format(k)+\
+                      "given in {0}".format(input_params_helper.INPUT_PARAMS_FILE)
+                  self.logger.error(msg)
+                  raise DjerbaPluginError(msg)
       if wrapper.my_param_is_null(constants.TUMOUR_ID):
           if self.workspace.has_file(core_constants.DEFAULT_SAMPLE_INFO):
               data = self.workspace.read_json(core_constants.DEFAULT_SAMPLE_INFO)
               tumour_id = data['tumour_id']
               wrapper.set_my_param(constants.TUMOUR_ID, tumour_id)
+              self.logger.debug("Read tumour ID from default sample info JSON")
           else:
               msg = "Cannot find tumour ID; must be manually specified or "+\
                   "given in {0}".format(core_constants.DEFAULT_SAMPLE_INFO)
               self.logger.error(msg)
               raise DjerbaPluginError(msg)
-
       # Get files for MSI, ctDNA
       donor = config[self.identifier][constants.DONOR]
       if wrapper.my_param_is_null(constants.MSI_FILE):
           wrapper.set_my_param(constants.MSI_FILE, self.get_file(donor, self.MSI_WORKFLOW, self.MSI_RESULTS_SUFFIX))
       if wrapper.my_param_is_null(constants.CTDNA_FILE):
           wrapper.set_my_param(constants.CTDNA_FILE, self.get_file(donor, self.CTDNA_WORKFLOW, self.CTDNA_RESULTS_SUFFIX))
-
       return config
 
     def extract(self, config):
@@ -142,7 +122,7 @@ class main(plugin_base):
 
       # Get parameters from config 
       tumour_id = wrapper.get_my_string(constants.TUMOUR_ID)
-      oncotree_code = wrapper.get_my_string(constants.ONCOTREE_CODE)
+      oncotree_code = wrapper.get_my_string(oncokb_constants.ONCOTREE_CODE)
       tcga_code = wrapper.get_my_string(constants.TCGA_CODE).lower() # tcga_code is always lowercase
       purity = wrapper.get_my_string(constants.PURITY_INPUT)
 
@@ -164,9 +144,8 @@ class main(plugin_base):
       results[constants.PURITY] = float(purity)*100
 
       # Annotate genomic biomarkers for therapy info/merge inputs
-      self.build_genomic_biomarkers(work_dir, oncotree_code, tumour_id)
-      merge_inputs = self.get_merge_inputs(work_dir)
-      
+      annotated_maf = self.annotate(work_dir, wrapper)
+      merge_inputs = self.get_merge_inputs(annotated_maf)
       data = {
           'plugin_name': 'Genomic Landscape and Biomarkers (TMB, MSI)',
           'version': self.PLUGIN_VERSION,
@@ -175,7 +154,6 @@ class main(plugin_base):
           'merge_inputs': merge_inputs,
           'results': results
       }
-
       return data
 
     def render(self, data):
@@ -194,26 +172,17 @@ class main(plugin_base):
           raise RuntimeError(msg) from err
       return results_path
 
-    def build_genomic_biomarkers(self, work_dir, oncotree_code, tumour_ID):
+    def annotate(self, work_dir, config):
         """
         # Writes and annotates the biomarkers file in the report directory
         """
         self.logger.debug("Annotating Genomic Biomarkers")
-        
-        # Get input file (to be annotated)
-        genomic_biomarkers_path = os.path.join(work_dir, constants.GENOMIC_BIOMARKERS)
-        # Get output file (after annotation)
-        genomic_biomarkers_annotated = os.path.join(work_dir, constants.GENOMIC_BIOMARKERS_ANNOTATED)
-        # Annotate the input file to the output file 
-        oncokb_annotator(
-            tumour_ID,
-            oncotree_code.upper(),
-            work_dir,
-            cache_params = None,
-            log_level=self.log_level,
-            log_path=self.log_path
-        ).annotate_biomarkers_maf(genomic_biomarkers_path, genomic_biomarkers_annotated)
-        # Return data for the biomarkers section of the output JSON
+        input_path = os.path.join(work_dir, constants.GENOMIC_BIOMARKERS)
+        output_path = os.path.join(work_dir, constants.GENOMIC_BIOMARKERS_ANNOTATED)
+        factory = annotator_factory(self.log_level, self.log_path)
+        annotator = factory.get_annotator(work_dir, config)
+        annotator.annotate_biomarkers_maf(input_path, output_path)
+        return output_path
          
     def make_biomarkers_maf(self, work_dir):
         maf_header = '\t'.join(["HUGO_SYMBOL","SAMPLE_ID","ALTERATION"])
@@ -224,7 +193,7 @@ class main(plugin_base):
             print(maf_header, file=genomic_biomarkers_file)
         return(genomic_biomarkers_path)
 
-    def get_merge_inputs(self, work_dir):
+    def get_merge_inputs(self, annotated_maf_path):
         """
         Read therapy information for merge inputs
         This is derived from the annotated biomarkers file.
@@ -233,8 +202,7 @@ class main(plugin_base):
         # read the tab-delimited input file
         treatments = []
         treatment_option_factory = tom_factory(self.log_level, self.log_path)
-        input_name = constants.GENOMIC_BIOMARKERS_ANNOTATED
-        with open(os.path.join(work_dir, input_name)) as input_file:
+        with open(annotated_maf_path) as input_file:
             reader = csv.DictReader(input_file, delimiter="\t")
             for row_input in reader:
                 # record therapy for all actionable alterations (OncoKB level 4 or higher)